Extract audio/video tooltip functionality to own component (BL-11107)
